feat: add new service for get id of product in API
